+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - If more than one `prototool.yaml` is found for the input directory or files,
   an error is returned.
 - The `prototool` binary package is moved under `internal`.
+- Remove the `protoc-url` global flag.
 
 
 ## [0.4.0] - 2018-06-22